+++ src/sll/api/_generated_raw.c
@@ -5077,6 +5077,28 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_thread_create_lock_raw(s
 
 
 
+__SLL_API_TYPE_sll_api_thread_create_semaphore sll_api_thread_create_semaphore(__SLL_API_ARGS_sll_api_thread_create_semaphore);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_thread_create_semaphore_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_object_t* a=NULL;
+	if (all>0){
+		a=*(al+0);
+		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_INT){
+			SLL_ACQUIRE(a);
+		}
+		else{
+			a=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_INT]);
+		}
+	}
+	else{
+		a=SLL_ACQUIRE_STATIC_INT(0);
+	}
+	sll_integer_t out=sll_api_thread_create_semaphore(a->dt.i);
+	SLL_RELEASE(a);
+	return SLL_FROM_INT(out);
+}
+
+
+
 __SLL_API_TYPE_sll_api_thread_release_lock sll_api_thread_release_lock(__SLL_API_ARGS_sll_api_thread_release_lock);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_thread_release_lock_raw(sll_object_t*const* al,sll_arg_count_t all){
 	sll_object_t* a=NULL;
@@ -5093,6 +5115,29 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_thread_release_lock_raw(
 		a=SLL_ACQUIRE_STATIC_INT(0);
 	}
 	sll_bool_t out=sll_api_thread_release_lock(a->dt.i);
+	SLL_RELEASE(a);
+	SLL_ACQUIRE(sll_static_int[out]);
+	return sll_static_int[out];
+}
+
+
+
+__SLL_API_TYPE_sll_api_thread_release_semaphore sll_api_thread_release_semaphore(__SLL_API_ARGS_sll_api_thread_release_semaphore);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_thread_release_semaphore_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_object_t* a=NULL;
+	if (all>0){
+		a=*(al+0);
+		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_INT){
+			SLL_ACQUIRE(a);
+		}
+		else{
+			a=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_INT]);
+		}
+	}
+	else{
+		a=SLL_ACQUIRE_STATIC_INT(0);
+	}
+	sll_bool_t out=sll_api_thread_release_semaphore(a->dt.i);
 	SLL_RELEASE(a);
 	SLL_ACQUIRE(sll_static_int[out]);
 	return sll_static_int[out];
@@ -5802,8 +5847,16 @@ static const internal_function_t _ifunc_data_ptr[]={
 		sll_api_thread_create_lock_raw
 	},
 	{
+		SLL_CHAR("sll:thread_create_semaphore"),
+		sll_api_thread_create_semaphore_raw
+	},
+	{
 		SLL_CHAR("sll:thread_release_lock"),
 		sll_api_thread_release_lock_raw
+	},
+	{
+		SLL_CHAR("sll:thread_release_semaphore"),
+		sll_api_thread_release_semaphore_raw
 	},
 	{
 		SLL_CHAR("sll:time_current"),
@@ -5825,5 +5878,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=165;
+const sll_function_index_t _ifunc_size=167;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);