+++ src/sll/include/sll/api/thread.h
@@ -38,6 +38,18 @@ __API_FUNC_DECL(thread_create_lock);
 
 /**
  * \flags api func
+ * \name sll_api_thread_create_semaphore
+ * \group thread-api
+ * \desc Docs!
+ * \arg I
+ * \ret I
+ */
+__API_FUNC_DECL(thread_create_semaphore);
+
+
+
+/**
+ * \flags api func
  * \name sll_api_thread_release_lock
  * \group thread-api
  * \desc Docs!
@@ -45,6 +57,18 @@ __API_FUNC_DECL(thread_create_lock);
  * \ret B
  */
 __API_FUNC_DECL(thread_release_lock);
+
+
+
+/**
+ * \flags api func
+ * \name sll_api_thread_release_semaphore
+ * \group thread-api
+ * \desc Docs!
+ * \arg I
+ * \ret B
+ */
+__API_FUNC_DECL(thread_release_semaphore);