+++ src/sll/include/sll/assembly.h
@@ -1220,25 +1220,25 @@
 
 /**
  * \flags macro var
- * \name SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_WAIT
- * \group assembly
- * \subgroup assembly-instruction
- * \desc Docs!
- * \type sll_assembly_instruction_type_t
- */
-#define SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_WAIT 100
-
-
-
-/**
- * \flags macro var
  * \name SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_ID
  * \group assembly
  * \subgroup assembly-instruction
  * \desc Docs!
  * \type sll_assembly_instruction_type_t
  */
-#define SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_ID 101
+#define SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_ID 100
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_WAIT
+ * \group assembly
+ * \subgroup assembly-instruction
+ * \desc Docs!
+ * \type sll_assembly_instruction_type_t
+ */
+#define SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_WAIT 101
 
 
 
@@ -1256,13 +1256,25 @@
 
 /**
  * \flags macro var
+ * \name SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_SEMAPHORE
+ * \group assembly
+ * \subgroup assembly-instruction
+ * \desc Docs!
+ * \type sll_assembly_instruction_type_t
+ */
+#define SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_SEMAPHORE 103
+
+
+
+/**
+ * \flags macro var
  * \name SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK
  * \group assembly
  * \subgroup assembly-instruction
  * \desc Docs!
  * \type sll_assembly_instruction_type_t
  */
-#define SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK 103
+#define SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK 104