+++ src/sll/io/generate_code.c
@@ -710,14 +710,17 @@ static const sll_node_t* _generate_code_internal(const code_generation_data_t* c
 				sll_file_write_char(wf,'}');
 				return o;
 			}
-		case SLL_NODE_TYPE_THREAD_WAIT:
-			PRINT_STATIC_STRING("!<<",wf);
-			break;
 		case SLL_NODE_TYPE_THREAD_ID:
 			PRINT_STATIC_STRING("!.",wf);
 			break;
+		case SLL_NODE_TYPE_THREAD_WAIT:
+			PRINT_STATIC_STRING("!<<",wf);
+			break;
 		case SLL_NODE_TYPE_THREAD_LOCK:
 			PRINT_STATIC_STRING("!<*",wf);
+			break;
+		case SLL_NODE_TYPE_THREAD_SEMAPHORE:
+			PRINT_STATIC_STRING("!<+",wf);
 			break;
 		default:
 			SLL_UNREACHABLE();