+++ src/sll/print.c
@@ -465,14 +465,17 @@ static const sll_node_t* _print_node_internal(const sll_source_file_t* sf,const 
 				sll_file_write_char(wf,'}');
 				return o;
 			}
-		case SLL_NODE_TYPE_THREAD_WAIT:
-			PRINT_STATIC_STRING("!<<",wf);
-			break;
 		case SLL_NODE_TYPE_THREAD_ID:
 			PRINT_STATIC_STRING("!.",wf);
 			break;
+		case SLL_NODE_TYPE_THREAD_WAIT:
+			PRINT_STATIC_STRING("!<<",wf);
+			break;
 		case SLL_NODE_TYPE_THREAD_LOCK:
 			PRINT_STATIC_STRING("!<*",wf);
+			break;
+		case SLL_NODE_TYPE_THREAD_SEMAPHORE:
+			PRINT_STATIC_STRING("!<+",wf);
 			break;
 		default:
 			SLL_UNREACHABLE();
@@ -1122,6 +1125,9 @@ __SLL_EXTERNAL void sll_print_assembly(const sll_assembly_data_t* a_dt,sll_file_
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_LOCK:
 				PRINT_STATIC_STRING("THREAD_LOCK",wf);
+				break;
+			case SLL_ASSEMBLY_INSTRUCTION_TYPE_THREAD_SEMAPHORE:
+				PRINT_STATIC_STRING("THREAD_SEMAPHORE",wf);
 				break;
 			default:
 				SLL_UNREACHABLE();